+++ Assets/NatShare/Plugins/Managed/INatShare.cs
@@ -8,9 +8,8 @@ namespace NatShareU {
 	using UnityEngine;
 
 	public interface INatShare {
-		bool ShareText (string text, ShareCallback callback);
-		bool ShareImage (byte[] pngData, string message, ShareCallback callback);
-		bool ShareMedia (string path, string message, ShareCallback callback);
+		bool Share (byte[] pngData, ShareCallback callback);
+		bool Share (string media, ShareCallback callback);
 		bool SaveToCameraRoll (byte[] pngData, string album);
         bool SaveToCameraRoll (string path, string album);
         Texture2D GetThumbnail (string videoPath, float time);